Split kol-card in shadow and WC component
Refs: #7869
